Bigger map, more filtering options, more data
Get rid of top navbar, make map full screen wide and much taller.
Add toggle to control cropping of data to visible area.
Restructure code that reads fro the request and creates the filters.
Exclude data that was collected using `MediaRecorder.AudioSource.VOICE_RECOGNITION`.
Show timestamp and proximity in infobox when clicking on a point on the map.